+++ tests/test_log.py
@@ -1,5 +1,5 @@
 from unittest.mock import patch
-
+import pytest
 from pynonymizer.log import get_logger, get_default_logger
 
 
@@ -20,3 +20,14 @@ def test_multipart_name(get_logger_mock):
     logger = get_logger(["a", "b", "cd", "ef", "ghi"])
     get_logger_mock.assert_called_once_with("pynonymizer.a.b.cd.ef.ghi")
 
+
+@patch("logging.getLogger")
+def test_no_logger_name_array(get_logger_mock):
+    with pytest.raises(ValueError):
+        logger = get_logger([])
+
+
+@patch("logging.getLogger")
+def test_no_logger_name_str(get_logger_mock):
+    with pytest.raises(ValueError):
+        logger = get_logger("")